+++ wasatch/WrapperWorker.py
@@ -1,4 +1,5 @@
 import threading
+import platform
 import logging
 import time
 from queue import Queue
@@ -13,6 +14,9 @@ from .SPIDevice            import SPIDevice
 from .BLEDevice            import BLEDevice
 from .TCPDevice            import TCPDevice
 from .Reading              import Reading
+
+if "windows" in platform.platform().lower():
+    from .IDSDevice        import IDSDevice
 
 log = logging.getLogger(__name__)
 
@@ -50,6 +54,7 @@ class WrapperWorker(threading.Thread):
             is_spi,
             is_ble,
             is_tcp,
+            is_ids,
             log_level,
             callback=None,
             alert_queue=None):
@@ -62,6 +67,7 @@ class WrapperWorker(threading.Thread):
         self.is_spi         = is_spi
         self.is_ble         = is_ble
         self.is_tcp         = is_tcp
+        self.is_ids         = is_ids
         self.command_queue  = command_queue
         self.response_queue = response_queue
         self.settings_queue = settings_queue
@@ -85,8 +91,13 @@ class WrapperWorker(threading.Thread):
     # one of the three queues (cmd inputs, response outputs, and
     # a one-shot SpectrometerSettings).
     def run(self):
-        is_options = (self.is_ocean, self.is_andor, self.is_ble, self.is_spi, self.is_tcp)
-        device_classes = (OceanDevice, AndorDevice, BLEDevice, SPIDevice, TCPDevice, WasatchDevice)
+        if "windows" in platform.platform().lower():
+            is_options = (self.is_ocean, self.is_andor, self.is_ble, self.is_spi, self.is_tcp, self.is_ids)
+            device_classes = (OceanDevice, AndorDevice, BLEDevice, SPIDevice, TCPDevice, IDSDevice, WasatchDevice)
+        else:
+            is_options = (self.is_ocean, self.is_andor, self.is_ble, self.is_spi, self.is_tcp)
+            device_classes = (OceanDevice, AndorDevice, BLEDevice, SPIDevice, TCPDevice, WasatchDevice)
+
         try:
             if any(is_options):
                 type_connection = is_options.index(True)